Switched compliance report distribution to explicit copy from masterfiles
Prior to this change the directory where we expect to import compliance report
definitions from (/var/cfengine/imports/compliance-reports) was populated with
data from the policy inputs. Since these assets live in the .no-distrib
directory they are not distributed to inputs by default and they need to be
explicitly copied down if desired (that's what this change is about).

Ticket: ENT-8333
Changelog: Title